test(azdo): split PR surface tests
Keep Azure DevOps provider tests under the file-size gate.

Refs #465
